requirements.txt
pycryptodome==3.13.0
pycryptodomex==3.11.0
scapy==2.6.0
